chore: refactor tests for campaign dispatcher (RMCCX-7619)
