+++ Tests/Tests/ReadyCampaignDispatcherSpec.swift
@@ -20,8 +20,12 @@ class ReadyCampaignDispatcherSpec: QuickSpec {
             var delegate: Delegate!
             var router: RouterMock!
             var httpSession: URLSessionMock!
+            var imageDetails: [String: ImageDetails]!
+            var images: [UIImage?]!
+            var result: [CarouselData]!
 
             beforeEach {
+                URLCache.shared.removeAllCachedResponses()
                 permissionService = DisplayPermissionServiceMock()
                 campaignRepository = CampaignRepositoryMock()
                 router = RouterMock()
@@ -45,6 +49,111 @@ class ReadyCampaignDispatcherSpec: QuickSpec {
 
             afterEach {
                 URLSessionMock.stopMockingURLSession()
+            }
+
+            context("when the carousel image URL is invalid") {
+                it("returns nil and does not cache the response") {
+                    let invalidURL = "invalid-url"
+
+                    dispatcher.fetchCarouselImage(for: invalidURL) { image in
+                        expect(image).to(beNil())
+
+                        let cachedResponse = URLCache.shared.cachedResponse(for: URLRequest(url: URL(string: invalidURL)!))
+                        expect(cachedResponse).to(beNil())
+                    }
+                }
+            }
+
+            context("when the network request fails for carousel image fetch") {
+                it("returns nil and does not cache the response") {
+                    let failedURL = "https://example.com/invalid-image.jpg"
+                    let url = URL(string: failedURL)!
+
+                    httpSession.httpResponse = HTTPURLResponse(url: url, statusCode: 404, httpVersion: nil, headerFields: nil)
+                    httpSession.responseData = nil
+                    httpSession.responseError = NSError(domain: "NetworkError", code: 404, userInfo: nil)
+
+                    dispatcher.fetchCarouselImage(for: failedURL) { image in
+                        expect(image).to(beNil())
+
+                        let cachedResponse = URLCache.shared.cachedResponse(for: URLRequest(url: url))
+                        expect(cachedResponse).to(beNil())
+                    }
+                }
+            }
+
+            context("when a valid image URL is provided in the carousel data") {
+                let validURL = "https://example.com/valid-image.jpg"
+                let url = URL(string: validURL)!
+
+                beforeEach {
+                    let imageData = UIImage(named: "test-image", in: .unitTests, with: nil)!.pngData()!
+                    httpSession.httpResponse = HTTPURLResponse(url: url, statusCode: 200, httpVersion: nil, headerFields: nil)
+                    httpSession.responseData = imageData
+                }
+
+                it("fetches and caches the image") {
+                    dispatcher.fetchCarouselImage(for: validURL) { image in
+                        expect(image).toNot(beNil())
+                        expect(image?.pngData()).to(equal(httpSession.responseData))
+
+                        let cachedResponse = URLCache.shared.cachedResponse(for: URLRequest(url: url))
+                        expect(cachedResponse).toNot(beNil())
+                        expect(cachedResponse?.data).to(equal(httpSession.responseData))
+                    }
+                }
+
+                it("returns the cached image for subsequent requests") {
+                    // Fetch the image once to cache it
+                    dispatcher.fetchCarouselImage(for: validURL) { _ in }
+
+                    // Fetch the image again to verify it comes from the cache
+                    dispatcher.fetchCarouselImage(for: validURL) { image in
+                        expect(image).toNot(beNil())
+                        let cachedResponse = URLCache.shared.cachedResponse(for: URLRequest(url: url))
+                        expect(cachedResponse).toNot(beNil())
+                        expect(image?.pngData()).to(equal(cachedResponse?.data))
+                    }
+                }
+            }
+
+            context("when the carousel data has no image URLs") {
+
+                it("returns an empty array") {
+                    let carousel = Carousel(images: [:])
+
+                    waitUntil { done in
+                        dispatcher.fetchImagesArray(from: carousel) { images in
+                            expect(images).to(beEmpty())
+                            done()
+                        }
+                    }
+                }
+            }
+
+            context("when the carousel data has invalid image URLs") {
+                beforeEach {
+                    httpSession.httpResponse = HTTPURLResponse(url: URL(string: "https://invalid-url.com")!,
+                                                               statusCode: 404,
+                                                               httpVersion: nil,
+                                                               headerFields: nil)
+                    httpSession.responseData = nil
+                    httpSession.responseError = NSError(domain: NSURLErrorDomain, code: URLError.badURL.rawValue, userInfo: nil)
+                }
+
+                it("returns an array with nil values") {
+                    let imageDetails: [String: ImageDetails] = [
+                        "1": ImageDetails(imgUrl: "https://invalid-url.com", link: "", altText: "")
+                    ]
+                    let carousel = Carousel(images: imageDetails)
+                    waitUntil { done in
+                        dispatcher.fetchImagesArray(from: carousel) { images in
+                            expect(images.count).to(equal(imageDetails.count))
+                            expect(images).to(containElementSatisfying { $0 == nil })
+                            done()
+                        }
+                    }
+                }
             }
 
             context("before dispatching") {
@@ -203,7 +312,6 @@ class ReadyCampaignDispatcherSpec: QuickSpec {
                     }
 
                     it("will display campaign when carousel data is defined") {
-
                         // swiftlint:disable line_length
                         let campaign =
                         TestHelpers.generateCampaign(id: "test",
@@ -420,6 +528,155 @@ class ReadyCampaignDispatcherSpec: QuickSpec {
                     expect(dispatcher.isDispatching).toEventually(beTrue())
                     expect(router.displayedCampaignsCount).toEventually(equal(1))
                     expect(dispatcher.isDispatching).toEventually(beFalse())
+                }
+            }
+
+            context("when data and images are valid") {
+                beforeEach {
+                    imageDetails = [
+                        "1": ImageDetails(imgUrl: "https://www.sample.com/image1.png", link: "https://example.com/1", altText: "Alt 1"),
+                        "2": ImageDetails(imgUrl: "https://www.sample.com/image2.png", link: "https://example.com/2", altText: "Alt 2"),
+                        "3": ImageDetails(imgUrl: "https://www.sample.com/image3.png", link: "https://example.com/3", altText: "Alt 3")
+                    ]
+                    images = [UIImage(named: "test-image", in: .unitTests, with: nil),
+                              UIImage(named: "test-image", in: .unitTests, with: nil),
+                              UIImage(named: "test-image", in: .unitTests, with: nil)
+                    ]
+                    result = dispatcher.createCarouselDataList(from: imageDetails, using: images)
+                }
+
+                it("should return a list of correct CarouselData objects") {
+                    expect(result.count).to(equal(3))
+
+                    expect(result[0].altText).to(equal("Alt 1"))
+                    expect(result[0].link).to(equal("https://example.com/1"))
+                    expect(result[0].image).toNot(beNil())
+
+                    expect(result[1].altText).to(equal("Alt 2"))
+                    expect(result[1].link).to(equal("https://example.com/2"))
+                    expect(result[1].image).toNot(beNil())
+
+                    expect(result[2].altText).to(equal("Alt 3"))
+                    expect(result[2].link).to(equal("https://example.com/3"))
+                    expect(result[2].image).toNot(beNil())
+                }
+            }
+
+            context("when images have nil entries") {
+                beforeEach {
+                    imageDetails = [
+                        "1": ImageDetails(imgUrl: "https://www.sample.com/image1.png", link: "https://example.com/1", altText: "Alt 1"),
+                        "2": ImageDetails(imgUrl: "https://www.sample.com/image2.png", link: "https://example.com/2", altText: "Alt 2"),
+                        "3": ImageDetails(imgUrl: "https://www.sample.com/image3.png", link: "https://example.com/3", altText: "Alt 3")
+                    ]
+                    images = [UIImage(named: "test-image", in: .unitTests, with: nil), nil]
+                    result = dispatcher.createCarouselDataList(from: imageDetails, using: images)
+                }
+
+                it("should handle nil images and still create data objects") {
+                    expect(result.count).to(equal(2))
+
+                    expect(result[0].image).toNot(beNil())
+                    expect(result[0].altText).to(equal("Alt 1"))
+                    expect(result[1].image).to(beNil())
+                    expect(result[1].altText).to(equal("Alt 2"))
+                }
+            }
+
+            context("when images count is fewer than data entries") {
+                beforeEach {
+                    imageDetails = [
+                        "1": ImageDetails(imgUrl: "https://www.sample.com/image1.png", link: "https://example.com/1", altText: "Alt 1"),
+                        "2": ImageDetails(imgUrl: "https://www.sample.com/image2.png", link: "https://example.com/2", altText: "Alt 2"),
+                        "3": ImageDetails(imgUrl: "https://www.sample.com/image3.png", link: "https://example.com/3", altText: "Alt 3")
+                    ]
+                    images = [UIImage(named: "test-image", in: .unitTests, with: nil),
+                              UIImage(named: "test-image", in: .unitTests, with: nil)]
+
+                    result = dispatcher.createCarouselDataList(from: imageDetails, using: images)
+                }
+
+                it("should only include entries with matching images") {
+                    expect(result.count).to(equal(2))
+
+                    expect(result[0].altText).to(equal("Alt 1"))
+                    expect(result[0].image).toNot(beNil())
+
+                    expect(result[1].altText).to(equal("Alt 2"))
+                    expect(result[1].image).toNot(beNil())
+                }
+            }
+
+            context("when data is empty") {
+                beforeEach {
+                    imageDetails = [:]
+                    images = [UIImage(named: "test-image", in: .unitTests, with: nil),
+                              UIImage(named: "test-image", in: .unitTests, with: nil)]
+                    result = dispatcher.createCarouselDataList(from: imageDetails, using: images)
+                }
+
+                it("should return an empty list") {
+                    expect(result).to(beEmpty())
+                }
+            }
+
+            context("when images are empty") {
+                beforeEach {
+                    imageDetails = [
+                        "1": ImageDetails(imgUrl: "https://www.sample.com/image1.png", link: "https://example.com/1", altText: "Alt 1"),
+                        "2": ImageDetails(imgUrl: "https://www.sample.com/image2.png", link: "https://example.com/2", altText: "Alt 2"),
+                        "3": ImageDetails(imgUrl: "https://www.sample.com/image3.png", link: "https://example.com/3", altText: "Alt 3")
+                    ]
+                    images = []
+
+                    result = dispatcher.createCarouselDataList(from: imageDetails, using: images)
+                }
+
+                it("should return an empty list") {
+                    expect(result).to(beEmpty())
+                }
+            }
+
+            context("when all images are nil") {
+                beforeEach {
+                    imageDetails = [
+                        "1": ImageDetails(imgUrl: "https://www.sample.com/image1.png", link: "https://example.com/1", altText: "Alt 1"),
+                        "2": ImageDetails(imgUrl: "https://www.sample.com/image2.png", link: "https://example.com/2", altText: "Alt 2")
+                    ]
+                    images = [nil, nil]
+                    result = dispatcher.createCarouselDataList(from: imageDetails, using: images)
+                }
+
+                it("should still create data objects with nil images") {
+                    expect(result.count).to(equal(2))
+
+                    expect(result[0].image).to(beNil())
+                    expect(result[0].altText).to(equal("Alt 1"))
+
+                    expect(result[1].image).to(beNil())
+                    expect(result[1].altText).to(equal("Alt 2"))
+                }
+            }
+
+            context("when keys are unsorted") {
+                beforeEach {
+                    imageDetails = [
+                        "3": ImageDetails(imgUrl: "https://www.sample.com/image3.png", link: "https://example.com/3", altText: "Alt 3"),
+                        "1": ImageDetails(imgUrl: "https://www.sample.com/image1.png", link: "https://example.com/1", altText: "Alt 1"),
+                        "2": ImageDetails(imgUrl: "https://www.sample.com/image2.png", link: "https://example.com/2", altText: "Alt 2")
+                    ]
+                    images = [UIImage(named: "test-image", in: .unitTests, with: nil),
+                              UIImage(named: "test-image", in: .unitTests, with: nil),
+                              UIImage(named: "test-image", in: .unitTests, with: nil)]
+                    result = dispatcher.createCarouselDataList(from: imageDetails, using: images)
+                }
+
+                it("should return sorted CarouselData list based on keys") {
+                    expect(result.count).to(equal(3))
+
+                    expect(result[0].altText).to(equal("Alt 1"))
+                    expect(result[1].altText).to(equal("Alt 2"))
+                    expect(result[2].altText).to(equal("Alt 3"))
                 }
             }
         }
